# Headers (HPUB) support