--- a/src/main/java/io/nats/client/Message.java
+++ b/src/main/java/io/nats/client/Message.java
@@ -36,6 +36,11 @@ public interface Message {
 	String getReplyTo();
 
 	/**
+	 * @return true if there are headers
+	 */
+	boolean hasHeaders();
+
+	/**
 	 * @return the headers object the message
 	 */
 	Headers getHeaders();