--- a/src/main/java/io/nats/client/impl/ByteArrayBuilder.java
+++ b/src/main/java/io/nats/client/impl/ByteArrayBuilder.java
@@ -1,4 +1,4 @@
-// Copyright 2015-2020 The NATS Authors
+// Copyright 2020 The NATS Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: