File by file notes:
<pre>
main/java/io/nats/client/
    Connection.java - interface change to support publish and request with headers
    Message.java    - interface change to headers
    Options.java    - changes to suport headers in the connect string
	
main/java/io/nats/client/impl/
    ByteArrayBuilder.java     - (BAB) utility class to simplying using ByteBuffers
    Headers.java              - object to manage the header map 
    NatsConnection.java       - allow publish and request with headers. Use BAB to build simple protocol messages. 
    NatsConnectionReader.java - read HMSG
    NatsConnectionWriter.java - write HPUB
    NatsMessage.java          - Headers support. constructor cleanup. builder for normal messages (ctor is fine for incoming and protocol)

examples/java/io/nats/examples/
    ExampleArgs.java - utility to read command line args to support header inputs
    NatsPub.java     - support headers, refactor to use example args, additional output
    NatsReply.java   - "
    NatsReq.java     - "
    NatsSub.java     - "

test/java/io/nats/client/
    NKeyTests.java       - extract utf-8 test strings to data file 
    OptionsTests.java    - connect protocol has "headers": true
    PublishTests.java    - publish api takes headers
    SubscriberTests.java - extract utf-8 test strings to data file 
 
test/java/io/nats/client/impl/
    ByteArrayBuilderTests.java - test BAB utility that it auto expands and the data is correct
    HeadersTests.java          - test the headers object
    MessageProtocolCreationBenchmark.java - NatsMessage no longer contstructs with CharBuffer, takes String or byte[] 
    MessageQueueBenchmark.java            - " 
    MessageQueueTests.java                - "
    NatsMessageTests.java                 - "

test/java/io/nats/client/utils/
    ResourceUtils.java - utility with method to read a text file from the resource folder

test/resources/
    utf8-test-strings.txt - test data
</pre>