@jameshilliard 2020-11-18
Rebased #364 on top of latest changes here.