@jameshilliard 2020-11-18
I did some more bench-marking to see how much of a performance impact the extra allocations created by the use of `ByteArrayBuilder` have overall, they seem to have a significant overall negative performance impact(in some cases around 50% slower) for some workloads while others have nearly no impact.

With #364:
> Task :NatsAutoBench.main()
Connecting to NATS server at nats://localhost:4222
Running warmup
Current memory usage is 990.90 mb / 1.02 gb / 16.00 gb free/total/max
Executing tests ........................................

PubOnly 0            10,000,000          4,899,764 msg/s       0.00 b/s
PubOnly 8            10,000,000          4,762,806 msg/s     36.34 mb/s
PubOnly 32           10,000,000          4,510,598 msg/s    137.65 mb/s
PubOnly 256          10,000,000          4,425,957 msg/s      1.06 gb/s
PubOnly 512          10,000,000          4,073,548 msg/s      1.94 gb/s
PubOnly 1024          1,000,000          3,019,998 msg/s      2.88 gb/s
PubOnly 4096            500,000            899,662 msg/s      3.43 gb/s
PubOnly 8192            100,000            464,286 msg/s      3.54 gb/s

PubSub 0             10,000,000          1,425,461 msg/s       0.00 b/s
PubSub 8             10,000,000          1,461,065 msg/s     11.15 mb/s
PubSub 32            10,000,000          1,284,064 msg/s     39.19 mb/s
PubSub 256           10,000,000          1,088,432 msg/s    265.73 mb/s
PubSub 512           10,000,000          1,022,268 msg/s    499.15 mb/s
PubSub 1024           1,000,000            867,574 msg/s    847.24 mb/s
PubSub 4096             500,000            368,595 msg/s      1.41 gb/s
PubSub 8192             100,000            203,657 msg/s      1.55 gb/s

PubDispatch 0        10,000,000          1,690,515 msg/s       0.00 b/s
PubDispatch 8        10,000,000          1,405,512 msg/s     10.72 mb/s
PubDispatch 32       10,000,000          1,514,133 msg/s     46.21 mb/s
PubDispatch 256      10,000,000          1,203,895 msg/s    293.92 mb/s
PubDispatch 512      10,000,000          1,170,347 msg/s    571.46 mb/s
PubDispatch 1024      1,000,000          1,061,729 msg/s      1.01 gb/s
PubDispatch 4096        500,000            368,722 msg/s      1.41 gb/s
PubDispatch 8192        100,000            157,165 msg/s      1.20 gb/s

ReqReply 0               20,000              9,140 msg/s       0.00 b/s
ReqReply 8               20,000              9,287 msg/s     72.56 kb/s
ReqReply 32              10,000              8,840 msg/s    276.25 kb/s
ReqReply 256             10,000              8,744 msg/s      2.13 mb/s
ReqReply 512             10,000              9,962 msg/s      4.86 mb/s
ReqReply 1024            10,000              9,571 msg/s      9.35 mb/s
ReqReply 4096            10,000              9,553 msg/s     37.32 mb/s
ReqReply 8192            10,000              8,316 msg/s     64.97 mb/s

Latency 0     5,000     40 /  50.76 / 171    +/- 0.75  (microseconds)
Latency 8     5,000     40 /  50.36 / 208    +/- 0.75  (microseconds)
Latency 32    5,000     38 /  49.37 / 205    +/- 0.73  (microseconds)
Latency 256   5,000     40 /  49.80 / 253    +/- 0.74  (microseconds)
Latency 512   5,000     40 /  50.50 / 261    +/- 0.75  (microseconds)
Latency 1024  5,000     39 /  50.60 / 190    +/- 0.75  (microseconds)
Latency 4096  5,000     42 /  51.67 / 196    +/- 0.76  (microseconds)
Latency 8192  5,000     44 /  53.90 / 878    +/- 0.82  (microseconds)

Final memory usage is 67.86 mb / 80.00 mb / 16.00 gb free/total/max

BUILD SUCCESSFUL in 2m 25s
2 actionable tasks: 1 executed, 1 up-to-date
9:14:28 AM: Task execution finished 'NatsAutoBench.main()'.


Without #364:

> Task :NatsAutoBench.main()
Connecting to NATS server at nats://localhost:4222
Running warmup
Current memory usage is 980.49 mb / 1.02 gb / 16.00 gb free/total/max
Executing tests ........................................

PubOnly 0            10,000,000          4,391,731 msg/s       0.00 b/s
PubOnly 8            10,000,000          4,379,475 msg/s     33.41 mb/s
PubOnly 32           10,000,000          4,620,284 msg/s    141.00 mb/s
PubOnly 256          10,000,000          3,628,144 msg/s    885.78 mb/s
PubOnly 512          10,000,000          3,305,155 msg/s      1.58 gb/s
PubOnly 1024          1,000,000          1,996,703 msg/s      1.90 gb/s
PubOnly 4096            500,000            697,794 msg/s      2.66 gb/s
PubOnly 8192            100,000            396,452 msg/s      3.02 gb/s

PubSub 0             10,000,000          1,508,454 msg/s       0.00 b/s
PubSub 8             10,000,000          1,260,994 msg/s      9.62 mb/s
PubSub 32            10,000,000          1,236,728 msg/s     37.74 mb/s
PubSub 256           10,000,000          1,058,462 msg/s    258.41 mb/s
PubSub 512           10,000,000            966,510 msg/s    471.93 mb/s
PubSub 1024           1,000,000            718,667 msg/s    701.82 mb/s
PubSub 4096             500,000            268,319 msg/s      1.02 gb/s
PubSub 8192             100,000            128,888 msg/s   1006.94 mb/s

PubDispatch 0        10,000,000          1,482,108 msg/s       0.00 b/s
PubDispatch 8        10,000,000          1,488,553 msg/s     11.36 mb/s
PubDispatch 32       10,000,000          1,400,385 msg/s     42.74 mb/s
PubDispatch 256      10,000,000          1,128,562 msg/s    275.53 mb/s
PubDispatch 512      10,000,000            994,440 msg/s    485.57 mb/s
PubDispatch 1024      1,000,000            764,356 msg/s    746.44 mb/s
PubDispatch 4096        500,000            307,758 msg/s      1.17 gb/s
PubDispatch 8192        100,000            165,218 msg/s      1.26 gb/s

ReqReply 0               20,000              9,294 msg/s       0.00 b/s
ReqReply 8               20,000              9,351 msg/s     73.06 kb/s
ReqReply 32              10,000             10,109 msg/s    315.91 kb/s
ReqReply 256             10,000              9,859 msg/s      2.41 mb/s
ReqReply 512             10,000             10,017 msg/s      4.89 mb/s
ReqReply 1024            10,000              9,650 msg/s      9.42 mb/s
ReqReply 4096            10,000              9,789 msg/s     38.24 mb/s
ReqReply 8192            10,000              8,961 msg/s     70.01 mb/s

Latency 0     5,000     40 /  51.13 / 187    +/- 0.77  (microseconds)
Latency 8     5,000     42 /  53.88 / 268    +/- 0.82  (microseconds)
Latency 32    5,000     41 /  50.75 / 147    +/- 0.77  (microseconds)
Latency 256   5,000     39 /  51.31 / 171    +/- 0.77  (microseconds)
Latency 512   5,000     40 /  53.18 / 249    +/- 0.80  (microseconds)
Latency 1024  5,000     41 /  52.35 / 385    +/- 0.79  (microseconds)
Latency 4096  5,000     41 /  53.61 / 823    +/- 0.81  (microseconds)
Latency 8192  5,000     43 /  55.03 / 846    +/- 0.85  (microseconds)

Final memory usage is 67.79 mb / 80.00 mb / 16.00 gb free/total/max

BUILD SUCCESSFUL in 2m 35s
2 actionable tasks: 2 executed
9:17:59 AM: Task execution finished 'NatsAutoBench.main()'.