@jameshilliard 2020-11-23
@scottf I've rebased #364 on your latest changes, I think it came out significantly cleaner this time as I was able to adapt your `Token` class to do `ByteBuffer` slicing without significant logic changes.