@jameshilliard 2020-11-29
Potential fix for the `nats-server` install failure:
https://github.com/nats-io/nats.java/blob/fd9b4b318fffaecda0235d30aebb4f8b6d7fd661/.travis.yml#L15-L20
https://github.com/nats-io/nats.java/blob/fd9b4b318fffaecda0235d30aebb4f8b6d7fd661/.travis.yml#L53-L55